+++ src/main/java/com/epam/ta/reportportal/core/integration/plugin/handler/impl/IntegrationTypeHandlerImpl.java
@@ -32,7 +32,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 /**
  * Implementation of {@link IntegrationTypeHandler} that provides CRU (create, read, update)
- * operations for {@link IntegrationType} entity
+ * operations for {@link IntegrationType} entity.
  *
  * @author <a href="mailto:budaevqwerty@gmail.com">Ivan Budayeu</a>
  */
